requirements.txt
colorama==0.4.1
pyyaml==3.13